part4 bloglist 4.21 step 9

+++ part4/bloglist_4_15-4_22/controllers/blogs.js
@@ -42,8 +42,22 @@ blogsRouter.post('/', async(request, response) => {
 })
 
 blogsRouter.delete('/:id', async(request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()    
+    const token = getTokenFrom(request)  
+    const decodedToken = jwt.verify(token, process.env.SECRET) 
+
+    if (!token || !decodedToken.id) {    
+        return response.status(401).json({ error: 'token missing or invalid' }) 
+    }
+
+    const user = await User.findById(decodedToken.id)
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user.id.toString()){
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()    
+    }else{
+        return response.status(401).json({ error: 'the post that you try delete not belogs you' }) 
+    }
 })
 
 blogsRouter.put('/:id', async(request, response, next) => {
